EXPERIMENTO/PRUEBA2/requirements.txt
langchain
langchain-openai
pydantic
python-dotenv
streamlit